pyproject.toml
[project]
name = "reformatter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fireworks-ai>=0.19.19",
    "hypothesis>=6.140.2",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
]
